[#161336844] modifications on file structure
