+++ backend/functions.py
@@ -11,4 +11,3 @@ class Products(object):
         }
         products.append(productItem)
         return products
-
